Fixed Multiple Matrix Scrolling
There was a small issue with scrolling on the small matrix where on the
command to shift to multiple matrices the bits previously shifted to the
first MAX would go to the second MAX which would display characters from
the other matrix there. I fixed this by clearing the MAX's at the start
of the command. The next thing I will work on is spacing on the
characters that take up more than five columns and changing some of the
characters to look better.